requirements.txt
pyrogram
pyrofork==2.3.61
TgCrypto
pyromod==1.5
schedule
APscheduler
# --- For-Database ------------ #
pymongo
motor
dnspython
# --- For-Web-Response ------- #
aiohttp
python-dotenv
# --- For-Web-Response ------- #
pytz
requests
bs4
shortzy
aiofiles
asyncio
Flask==1.1.2
gunicorn==20.1.0
Jinja2==3.0.3
werkzeug==2.0.2
itsdangerous==2.0.1
